requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
altair==4.1.0
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.4.2
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
fsspec==0.8.5
future==0.18.2
gdown==3.12.2
gitdb==4.0.5
GitPython==3.1.12
google-auth==1.25.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
idna==2.10
importlib-metadata==3.4.0
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
nltk==3.5
notebook==6.2.0
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pandas==1.2.1
pandas-path==0.1.2
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.14.0
ptyprocess==0.7.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycparser==2.20
pydeck==0.6.0
Pygments==2.7.4
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-lightning==1.1.8
pytz==2021.1
PyYAML==5.3.1
pyzmq==22.0.2
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.0
Send2Trash==1.5.0
sentence-transformers==0.4.1.2
sentencepiece==0.1.95
six==1.15.0
smmap==3.0.5
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tokenizers==0.10.1
toml==0.10.2
toolz==0.11.1
torch==1.6.0
torchvision==0.7.0
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
transformers==4.3.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
validators==0.18.2
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.0
